+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - The index creation is now run on the `init` hook with a priority of *1000*. This allow various dependencies to execute before RediPress indexing.
+- Removed `post_object` and `permalink` from the index as they are never queried.
 
 ### Fixed
 - String escaping fixed for post data indexing.